first of many commits to restructure RSW code to look like heat-2d code -- serial numpy code is working
