+++ approach_2/setup_mpi.py
@@ -1,0 +1,243 @@
+import numpy as np
+from mpi4py import MPI
+from sadourny_setup import I, I_XP, I_XN, I_YP, I_YN, I_XP_YP, I_XP_YN, I_XN_YP, \
+                           euler, ab2, ab3
+from flux_ener_f2py77 import euler_f as ener_Euler_f77, \
+                             ab2_f as ener_AB2_f77,      \
+                             ab3_f as ener_AB3_f77,       \
+                             flux_ener as flux_ener_F77
+from flux_ener_f2py90 import euler_f as ener_Euler_f90, \
+                             ab2_f as ener_AB2_f90,      \
+                             ab3_f as ener_AB3_f90,       \
+                             flux_ener as flux_ener_F90
+comm = MPI.COMM_WORLD
+
+
+def solver_1D_helper(u, ranks, ghost_arr, tags, params):
+
+
+    f0, gp, H0 = params.consts
+
+    comm.Barrier()         # start MPI timer
+    t_start = MPI.Wtime()
+
+    # Euler step
+    uvh, NLnm, energy[0], enstr[0] = Flux_Euler(uvh, params, dims, rank, p, col, tags)
+    uvh = set_uvh_bdr(uvh, rank, p, nx, col, tags)
+
+    # AB2 step
+    uvh, NLn, energy[1], enstr[1]  = Flux_AB2(uvh, NLnm, params, dims, rank, p, col, tags)
+    uvh = set_uvh_bdr(uvh, rank, p, nx, col, tags)
+
+     # loop through time
+    for n in range(3, Nt):
+        # AB3 step
+        uvh, NL, energy[n-1], enstr[n-1] = Flux_AB3(uvh, NLn, NLnm, params, dims, rank, p, col, tags)
+
+        # Reset fluxes
+        NLnm, NLn = NLn, NL
+
+        uvh = set_uvh_bdr(uvh, rank, p, nx, col, tags)
+        UVHG = gather_uvh(uvh, uvhG, UVHG, rank, p, nx, Ny, n)
+
+    comm.Barrier()
+    t_total = (MPI.Wtime() - t_start)  # stop MPI timer
+
+    return t_total
+
+
+def flux_sw_ener_MPI_1D(uvh, params, rank, p, col, tags):
+    # All terms (h, U, V, B, etc...) are calculated in numpy.
+
+    # Define parameters
+    dx, dy     = params[0], params[1]
+    f0, gp, H0 = params[2], params[3], params[4]
+    Nx, Ny     = dims
+
+    # Pull out primitive variables
+    u, v, h = uvh[0, :, :], uvh[1, :, :],  H0 + uvh[2, :, :]
+
+    # Initialize fields
+    U, V = np.zeros((Ny+2, Nx+2)), np.zeros((Ny+2, Nx+2))
+    B, q = np.zeros((Ny+2, Nx+2)), np.zeros((Ny+2, Nx+2))
+    flux = np.zeros((3, Ny, Nx))
+
+    # Compute U, V, B, q
+    U[1:-1, 1:-1] = 0.5*(I(h) + I_XP(h)) * I(u)
+    V[1:-1, 1:-1] = 0.5*(I(h) + I_YP(h)) * I(v)
+    B[1:-1, 1:-1] = gp*I(h) + 0.25*(I(u)**2 + I_XN(u)**2 + I(v)**2 + I_YN(v)**2)
+    q[1:-1, 1:-1] = 4*((I_XP(v) - I(v)) / dx - (I_YP(u) - I(u)) / dy + f0) /   \
+                       (I(h) + I_YP(h) + I_XP(h) + I_XP_YP(h))
+
+    # Enforce BCs using MPI
+    U = set_mpi_bc(U, rank, p, col, tags)
+    V = set_mpi_bc(V, rank, p, col, tags)
+    B = set_mpi_bc(B, rank, p, col, tags)
+    q = set_mpi_bc(q, rank, p, col, tags)
+    # U = even(U)
+    # V = odd(V)
+    # B = even(B)
+    # q = even(q)
+
+    # Compute fluxes
+    flux[0, :, :] =  0.25*(I(q) * (I_XP(V) + I(V)) + I_YN(q) * (I_XP_YN(V) + I_YN(V))) - \
+                    (I_XP(B) - I(B))/dx
+    flux[1, :, :] = -0.25*(I(q) * (I_YP(U) + I(U)) + I_XN(q) * (I_XN_YP(U) + I_XN(U))) - \
+                    (I_YP(B) - I(B))/dy
+    flux[2, :, :] = -(U[1:-1, 1:-1] - U[1:-1, 0:-2])/dx - (V[1:-1, 1:-1] - V[0:-2, 1:-1])/dy
+
+    # compute energy and enstrophy
+    energy = 0.5*np.mean(gp*I(h)**2 + 0.5*I(h)*(I(u)**2 + I_XN(u)**2 + I(v)**2 + I_YN(v)**2))
+    enstrophy = 0.125*np.mean((I(h) + I_YP(h) + I_XP(h) + I_XP_YP(h)) * I(q)**2)
+
+    return flux, energy, enstrophy
+
+"""
+def flux_sw_ener_f90_MPI(uvh, params, dims, rank, p, col, tags):
+    # All terms (h, U, V, B, etc...) are calculated in F90.
+
+    # Define parameters
+    dx, dy     = params[0], params[1]
+    f0, gp, H0 = params[2], params[3], params[4]
+    Nx, Ny     = dims
+
+    # Pull out primitive variables
+    u, v, h = uvh[0, :, :], uvh[1, :, :],  H0 + uvh[2, :, :]
+
+    # Initialize fields
+    U, V = np.zeros((Ny+2, Nx+2)), np.zeros((Ny+2, Nx+2))
+    B, q = np.zeros((Ny+2, Nx+2)), np.zeros((Ny+2, Nx+2))
+    flux = np.zeros((3, Ny, Nx))
+
+    # Compute U, V, B, q
+    U[1:-1, 1:-1] = 0.5*(I(h) + I_XP(h)) * I(u)
+    V[1:-1, 1:-1] = 0.5*(I(h) + I_YP(h)) * I(v)
+    B[1:-1, 1:-1] = gp*I(h) + 0.25*(I(u)**2 + I_XN(u)**2 + I(v)**2 + I_YN(v)**2)
+    q[1:-1, 1:-1] = 4*((I_XP(v) - I(v)) / dx - (I_YP(u) - I(u)) / dy + f0) /   \
+                       (I(h) + I_YP(h) + I_XP(h) + I_XP_YP(h))
+
+    # Enforce BCs using MPI
+    U = set_mpi_bc(U, rank, p, col, tags)
+    V = set_mpi_bc(V, rank, p, col, tags)
+    B = set_mpi_bc(B, rank, p, col, tags)
+    q = set_mpi_bc(q, rank, p, col, tags)
+    # U = even(U)
+    # V = odd(V)
+    # B = even(B)
+    # q = even(q)
+
+    # Compute fluxes
+    flux[0, :, :] =  0.25*(I(q) * (I_XP(V) + I(V)) + I_YN(q) * (I_XP_YN(V) + I_YN(V))) - \
+                    (I_XP(B) - I(B))/dx
+    flux[1, :, :] = -0.25*(I(q) * (I_YP(U) + I(U)) + I_XN(q) * (I_XN_YP(U) + I_XN(U))) - \
+                    (I_YP(B) - I(B))/dy
+    flux[2, :, :] = -(U[1:-1, 1:-1] - U[1:-1, 0:-2])/dx - (V[1:-1, 1:-1] - V[0:-2, 1:-1])/dy
+
+    # compute energy and enstrophy
+    energy = 0.5*np.mean(gp*I(h)**2 + 0.5*I(h)*(I(u)**2 + I_XN(u)**2 + I(v)**2 + I_YN(v)**2))
+    enstrophy = 0.125*np.mean((I(h) + I_YP(h) + I_XP(h) + I_XP_YP(h)) * I(q)**2)
+
+    return flux, energy, enstrophy
+"""
+
+
+def ener_Euler_MPI(uvh, params, dims, rank, p, col, tags):
+    # MPI'd pure Numpy
+    NLnm, energy, enstr = flux_sw_ener_MPI(uvh, params, dims, rank, p, col, tags)
+    uvh[:, 1:-1, 1:-1]  = euler(uvh, params[5], NLnm)
+    return uvh, NLnm, energy, enstr
+
+
+def ener_AB2_MPI(uvh, NLnm, params, dims, rank, p, col, tags):
+    NLn, energy, enstr = flux_sw_ener_MPI(uvh, params, dims, rank, p, col, tags)
+    uvh[:, 1:-1, 1:-1] = ab2(uvh, params[5], NLn, NLnm)
+    return uvh, NLn, energy, enstr
+
+
+def ener_AB3_MPI(uvh, NLn, NLnm, params, dims, rank, p, col, tags):
+    NL, energy, enstr  = flux_sw_ener_MPI(uvh, params, dims, rank, p, col, tags)
+    uvh[:, 1:-1, 1:-1] = ab3(uvh, params[5], NL, NLn, NLnm)
+    return uvh, NL, energy, enstr
+
+
+def set_uvh_bdr(uvh, rank, p, nx, col, tags):
+
+    u, v, h = uvh[0, :, :], uvh[1, :, :], uvh[2, :, :]
+
+    u = set_mpi_bc(u, rank, p, col, tags)
+    v = set_mpi_bc(v, rank, p, col, tags)
+    h = set_mpi_bc(h, rank, p, col, tags)
+
+    # uvh = np.dstack([u, v, h])
+    uvh[0, :, :], uvh[1, :, :], uvh[2, :, :] = u, v, h
+
+    return uvh
+
+
+def create_global_objects(rank, xG, yG, Nx, Ny, N, dx, hmax, Lx):
+    if rank == 0:
+        xx, yy = np.meshgrid(xG, yG)
+        xuG = xx
+        xvG = xx - dx/2
+        xhG = xx - dx/2
+        yhG = yy
+
+        uvhG = np.zeros([3, Ny, Nx])                # global initial solution
+        UVHG = np.empty((3, Ny, Nx, N), dtype='d')  # set of ALL global solutions
+
+        uvhG[0, :, :] = 0*xuG
+        uvhG[1, :, :] = 0*xvG
+        uvhG[2, :, :] = hmax*np.exp(-(xhG**2 + yhG**2)/(Lx/6.0)**2)
+        UVHG[:, :, :, 0] = uvhG
+
+        uvhG = uvhG.flatten()
+
+    else:
+        uvhG = None
+        UVHG = None
+
+    return (uvhG, UVHG)
+
+
+def gather_uvh(uvh, uvhG, UVHG, rank, p, nx, Ny, n):
+    """
+    Gather each process' slice of uvh and store it in the global set of
+    solutions. Gathering is done on process 0.
+
+    Parameters
+    ----------
+    uvh : ndarray
+        process' slice of the solution (dimensions: (Ny, nx+2))
+    uvhG : ndarray
+        the global solution, excluding ghost points (dimensions: (Ny, p*nx))
+    UVHG : ndarray
+        the set of global solutions (dimensions: (Ny, p*nx, N)) where N is the
+        number of time steps defined elsewhere.
+    rank : int
+        rank of this process
+    p : int
+        number of processes
+    nx : int
+        the number of real (non-ghost) x points per process
+    Ny : int
+        the number of y points
+    n : int
+        time-step number
+
+    Returns
+    -------
+    ndarray or None
+        If this process' rank is 0, then the newest solution is added to UVHG
+        and UVHG is returned. Otherwise, None is returned.
+    """
+    comm.Gather(uvh[:, 1:-1, 1:-1].flatten(), uvhG, root=0)
+    if rank == 0:
+        # evenly split ug into a list of p parts
+        temp = np.array_split(uvhG, p)
+        # reshape each part
+        temp = [a.reshape(3, Ny, nx) for a in temp]
+
+        UVHG[:, :, :, n] = np.dstack(temp)
+        return UVHG
+    else:
+        return None
